Add controller view for the Authors page, test properties validation.
