+++ react.flux/src/components/authors/authorList.js
@@ -1,0 +1,38 @@
+'use strict';
+
+var createReactClass = require('create-react-class');
+var React = require('react');
+var PropTypes = require('prop-types');
+
+var AuthorList = createReactClass({
+    propTypes: {
+        authors: PropTypes.array.isRequired
+    },
+
+    render: function () {
+        var createAuthorRow = function (author) {
+            return (
+                <tr key={author.id}>
+                    <td><a href={"#/authors/" + author.id}>{author.id}</a></td>
+                    <td>{author.firstName} {author.lastName}</td>
+                </tr>
+            );
+        };
+
+        return (
+        <div>
+            <table className="table">
+                <thead>
+                    <th>ID</th>
+                    <th>Name</th>
+                </thead>
+                <tbody>
+                    {this.props.authors.map(createAuthorRow, this)}
+                </tbody>
+            </table>
+        </div>
+        );
+    }
+});
+
+module.exports = AuthorList;
