+++ react.flux/src/components/authors/authorPage.js
@@ -3,6 +3,7 @@
 var createReactClass = require('create-react-class');
 var React = require('react');
 var AuthorApi = require('../../api/authorApi');
+var AuthorList = require('./authorList');
 
 var Authors = createReactClass({
     getInitialState: function () {
@@ -11,32 +12,15 @@ var Authors = createReactClass({
         };
     },
 
-    componentWillMount: function () {
+    componentDidMount: function () {
         this.setState({ authors: AuthorApi.getAllAuthors( )});
     },
 
     render: function () {
-        var createAuthorRow = function (author) {
-            return (
-                <tr key={author.id}>
-                    <td><a href={"#/authors/" + author.id}>{author.id}</a></td>
-                    <td>{author.firstName} {author.lastName}</td>
-                </tr>
-            );
-        };
-
         return (
         <div>
             <h1>Authors</h1>
-            <table className="table">
-                <thead>
-                    <th>ID</th>
-                    <th>Name</th>
-                </thead>
-                <tbody>
-                    {this.state.authors.map(createAuthorRow, this)}
-                </tbody>
-            </table>
+            <AuthorList authors={this.state.authors} />
         </div>
         );
     }
